+++ bluetooth_low_energy_platform_interface/lib/src/event_args.dart
@@ -1,2 +1,2 @@
 /// The base event arguments.
-base class EventArgs {}
+abstract base class EventArgs {}