+++ bluetooth_low_energy_platform_interface/lib/src/peripheral_manager.dart
@@ -1,9 +1,8 @@
 import 'dart:typed_data';
 
-import 'package:plugin_platform_interface/plugin_platform_interface.dart';
-
 import 'advertisement.dart';
 import 'bluetooth_low_energy_manager.dart';
+import 'bluetooth_low_energy_plugin.dart';
 import 'central.dart';
 import 'connection_state.dart';
 import 'event_args.dart';
@@ -18,10 +17,7 @@ final class CentralConnectionStateChangedEventArgs extends EventArgs {
   final ConnectionState state;
 
   /// Constructs a [CentralConnectionStateChangedEventArgs].
-  CentralConnectionStateChangedEventArgs(
-    this.central,
-    this.state,
-  );
+  CentralConnectionStateChangedEventArgs(this.central, this.state);
 }
 
 /// The central MTU changed event arguments.
@@ -33,17 +29,11 @@ final class CentralMTUChangedEventArgs extends EventArgs {
   final int mtu;
 
   /// Constructs a [CentralMTUChangedEventArgs].
-  CentralMTUChangedEventArgs(
-    this.central,
-    this.mtu,
-  );
+  CentralMTUChangedEventArgs(this.central, this.mtu);
 }
 
 /// The GATT characteristic read requested event arguments.
 final class GATTCharacteristicReadRequestedEventArgs extends EventArgs {
-  /// The central which read this characteristic.
-  final Central central;
-
   /// The characteristic to read the value of.
   final GATTCharacteristic characteristic;
 
@@ -51,18 +41,11 @@ final class GATTCharacteristicReadRequestedEventArgs extends EventArgs {
   final GATTReadRequest request;
 
   /// Constructs a [GATTCharacteristicReadRequestedEventArgs].
-  GATTCharacteristicReadRequestedEventArgs(
-    this.central,
-    this.characteristic,
-    this.request,
-  );
+  GATTCharacteristicReadRequestedEventArgs(this.characteristic, this.request);
 }
 
 /// The GATT characteristic write requested event arguments.
 final class GATTCharacteristicWriteRequestedEventArgs extends EventArgs {
-  /// The central which wrote this characteristic.
-  final Central central;
-
   /// The characteristic to write the value of.
   final GATTCharacteristic characteristic;
 
@@ -70,18 +53,11 @@ final class GATTCharacteristicWriteRequestedEventArgs extends EventArgs {
   final GATTWriteRequest request;
 
   /// Constructs a [GATTCharacteristicWriteRequestedEventArgs].
-  GATTCharacteristicWriteRequestedEventArgs(
-    this.central,
-    this.characteristic,
-    this.request,
-  );
+  GATTCharacteristicWriteRequestedEventArgs(this.characteristic, this.request);
 }
 
 /// The GATT characteristic notify state changed event arguments.
 final class GATTCharacteristicNotifyStateChangedEventArgs extends EventArgs {
-  /// The central which set this notify state.
-  final Central central;
-
   /// The GATT characteristic which notify state changed.
   final GATTCharacteristic characteristic;
 
@@ -90,17 +66,11 @@ final class GATTCharacteristicNotifyStateChangedEventArgs extends EventArgs {
 
   /// Constructs a [GATTCharacteristicNotifyStateChangedEventArgs].
   GATTCharacteristicNotifyStateChangedEventArgs(
-    this.central,
-    this.characteristic,
-    this.state,
-  );
+      this.characteristic, this.state);
 }
 
 /// The GATT descriptor read requested event arguments.
 final class GATTDescriptorReadRequestedEventArgs extends EventArgs {
-  /// The central which read this descriptor.
-  final Central central;
-
   /// The descriptor to read the value of.
   final GATTDescriptor descriptor;
 
@@ -108,18 +78,11 @@ final class GATTDescriptorReadRequestedEventArgs extends EventArgs {
   final GATTReadRequest request;
 
   /// Constructs a [GATTDescriptorReadRequestedEventArgs].
-  GATTDescriptorReadRequestedEventArgs(
-    this.central,
-    this.descriptor,
-    this.request,
-  );
+  GATTDescriptorReadRequestedEventArgs(this.descriptor, this.request);
 }
 
 /// The GATT descriptor write requested event arguments.
 final class GATTDescriptorWriteRequestedEventArgs extends EventArgs {
-  /// The central which wrote this descriptor.
-  final Central central;
-
   /// The descriptor to write the value of.
   final GATTDescriptor descriptor;
 
@@ -127,26 +90,15 @@ final class GATTDescriptorWriteRequestedEventArgs extends EventArgs {
   final GATTWriteRequest request;
 
   /// Constructs a [GATTDescriptorWriteRequestedEventArgs].
-  GATTDescriptorWriteRequestedEventArgs(
-    this.central,
-    this.descriptor,
-    this.request,
-  );
+  GATTDescriptorWriteRequestedEventArgs(this.descriptor, this.request);
 }
 
 /// An object that manages and advertises peripheral services exposed by this app.
 abstract interface class PeripheralManager
     implements BluetoothLowEnergyManager {
-  static PeripheralManager? _instance;
-
   /// Gets the instance of [PeripheralManager] to use.
   factory PeripheralManager() {
-    final instance = PlatformPeripheralManager.instance;
-    if (instance != _instance) {
-      instance.initialize();
-      _instance = instance;
-    }
-    return instance;
+    return BluetoothLowEnergyPlugin.instance.createPeripheralManager();
   }
 
   /// Callback indicating when a remote device has been connected or disconnected.
@@ -210,8 +162,8 @@ abstract interface class PeripheralManager
   ///
   /// [includeTXPowerLevel] Whether the transmission power level should be included
   /// in the advertise packet. TX power level field takes 3 bytes in advertise
-  /// packet. This argument is available on Android and Windows, throws [UnsupportedError]
-  /// on other platforms.
+  /// packet. This argument is available on Android, throws [UnsupportedError] on
+  /// other platforms.
   Future<void> startAdvertising(
     Advertisement advertisement, {
     bool? includeDeviceName,
@@ -260,36 +212,6 @@ abstract interface class PeripheralManager
   Future<void> notifyCharacteristic(
     GATTCharacteristic characteristic, {
     required Uint8List value,
-    required Central central,
+    List<Central>? centrals,
   });
-}
-
-/// Platform-specific implementations should implement this class to support
-/// [PlatformPeripheralManager].
-abstract base class PlatformPeripheralManager
-    extends PlatformBluetoothLowEnergyManager implements PeripheralManager {
-  static final Object _token = Object();
-
-  static PlatformPeripheralManager? _instance;
-
-  /// The default instance of [PlatformPeripheralManager] to use.
-  static PlatformPeripheralManager get instance {
-    final instance = _instance;
-    if (instance == null) {
-      throw UnimplementedError(
-          'PeripheralManager is not implemented on this platform.');
-    }
-    return instance;
-  }
-
-  /// Platform-specific implementations should set this with their own
-  /// platform-specific class that extends [PlatformPeripheralManager] when
-  /// they register themselves.
-  static set instance(PlatformPeripheralManager instance) {
-    PlatformInterface.verifyToken(instance, _token);
-    _instance = instance;
-  }
-
-  /// Constructs a [PlatformPeripheralManager].
-  PlatformPeripheralManager() : super(token: _token);
 }